[#365] Move queueDeveloperAppResponse to ApigeeMockApiClientHelperTrait
