+++ tests/modules/apigee_mock_api_client/tests/src/Traits/ApigeeMockApiClientHelperTrait.php
@@ -23,6 +23,7 @@ use Apigee\Edge\Api\Management\Entity\Company;
 use Apigee\Edge\Api\Management\Entity\Organization;
 use Apigee\MockClient\Generator\ApigeeSdkEntitySource;
 use Drupal\apigee_edge\Entity\Developer;
+use Drupal\apigee_edge\Entity\DeveloperAppInterface;
 use Drupal\apigee_edge\Entity\DeveloperInterface;
 use Drupal\Tests\apigee_edge\Traits\ApigeeEdgeUtilTestTrait;
 use Drupal\user\UserInterface;
@@ -219,6 +220,29 @@ trait ApigeeMockApiClientHelperTrait {
     $context['developers'] = $developers;
 
     $this->stack->queueMockResponse(['developers_in_company' => $context]);
+  }
+
+  /**
+   * Add an app analytics mock response to the stack.
+   *
+   * @param \Drupal\apigee_edge\Entity\DeveloperAppInterface $app
+   *   The app.
+   * @param int $response_code
+   *   Response code, defaults to 200.
+   */
+  protected function queueDeveloperAppResponse(DeveloperAppInterface $app, $response_code = 200) {
+    $this->stack->queueMockResponse([
+      'get_developer_app' => [
+        'status_code' => $response_code,
+        'app' => [
+          'appId' => $app->getAppId() ?: $this->randomMachineName(),
+          'name' => $app->getName(),
+          'status' => $app->getStatus(),
+          'displayName' => $app->getStatus(),
+          'developerId' => $app->getDeveloperId(),
+        ],
+      ],
+    ]);
   }
 
   /**
